deepseek_ocr/requirements.txt
torch>=2.0.0
transformers>=4.40.0
tokenizers>=0.15.0
einops>=0.7.0
addict>=2.4.0
easydict>=1.10
matplotlib>=3.7.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pillow>=10.0.0
numpy>=1.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
accelerate>=0.24.0
flash-attn>=2.0.0; platform_system != "Darwin"

